requirements.txt
beir
pyyaml

sentence-transformers
rank_bm25
numpy
scipy
torch